Preparation for release
* Fixed bug where re-init of chibi stack doesn't initialize previous
sequence and previous source address. This led to some strange bugs
where packets would get flagged as duplicate packets and dropped.
* Added support for ATMega1284P in promiscuous mode using very large
receive buffer sizes. This is possible because the ATMega1284P has 16 kB
RAM as opposed to the ATMega328P with 2 kB RAM.
* Added support for Freaklabs FreakUSB, Arashi, and Saboten product
lines. These will be introduced shortly.
* Added support for OQPSK 1000 kb/s, 500 kb/s, and BPSK 20 kb/s data
rates as well as the standard 250 kb/s for the 900 MHz boards
* Added support for OQPSK 2000 kb/s, 1000 kb/s, and 500 kb/s as well as
standard 250 kb/s for the 2.4 GHz boards